changing the index.css & moving to styles folder
