--- frontend/src/index.css
@@ -1,85 +1,0 @@
-@import url('https://fonts.googleapis.com/css?family=Lato|Roboto&display=swap');
-
-html {
-  box-sizing: border-box;
-  font-size: 62.5%;
-}
-
-html,
-body {
-  height: 100%;
-}
-
-body {
-  padding: 0;
-  margin: 0;
-  background-color: rgb(207, 28, 28);
-  color: #333;
-  font-size: 1.6rem;
-  font-family: 'Lato', Arial, Helvetica, sans-serif;
-}
-
-*,
-*:before,
-*:after {
-  box-sizing: inherit;
-  -webkit-font-smoothing: antialiased;
-}
-
-h1,
-h2,
-h3,
-h4,
-h5,
-h6 {
-  font-family: 'Roboto', Arial, Helvetica, sans-serif;
-  font-weight: normal;
-  color: #333;
-  margin: 0;
-  font-size: 1.6rem;
-}
-
-p {
-  line-height: 1.6;
-  font-size: 1.6rem;
-}
-
-a,
-a:visited,
-a:active,
-a:hover {
-  color: inherit;
-  text-decoration: none;
-}
-
-.app-header {
-  width: 100%;
-  height: 10rem;
-  background: rgba(196, 196, 196, 0.2);
-  display: flex;
-  align-items: center;
-  padding: 0 4rem;
-}
-
-.app {
-  height: 100vh;
-}
-
-.app-footer {
-  position: fixed;
-  left: 0;
-  bottom: 0;
-  width: 100%;
-  background-color: rgb(185, 218, 200);
-  color: white;
-  text-align: center;
-  height: 80px;
-}
-
-.logo {
-  width: 31px;
-  height: 31px;
-  background: rgba(255, 255, 255, 0.2);
-  margin: 16px 24px 16px 0;
-  float: left;
-}
