+++ frontend/src/index1.css
@@ -1,0 +1,22 @@
+
+
+.app {
+  height: 100vh;
+}
+
+.app-footer {
+  position: fixed;
+  left: 0;
+  bottom: 0;
+  width: 100%;
+  text-align: center;
+  height: 80px;
+}
+
+.logo {
+  width: 31px;
+  height: 31px;
+  background: rgba(255, 255, 255, 0.2);
+  margin: 16px 24px 16px 0;
+  float: left;
+}
