+++ frontend/src/styles/index.css
@@ -1,0 +1,200 @@
+@import url('https://fonts.googleapis.com/css?family=Lato|Roboto&display=swap');
+@import url('https://fonts.googleapis.com/css?family=Istok+Web&display=swap');
+@import url('https://fonts.googleapis.com/css?family=Karla&display=swap');
+html {
+  box-sizing: border-box;
+  font-size: 62.5%;
+}
+
+html,
+body {
+  height: 100%;
+}
+
+body {
+  padding: 0;
+  margin: 0;
+  background-color: #fff;
+  color: #333;
+  font-size: 1.5rem;
+  font-family: 'Lato', Arial, Helvetica, sans-serif;
+}
+
+*,
+*:before,
+*:after {
+  box-sizing: inherit;
+  -webkit-font-smoothing: antialiased;
+}
+
+h1,
+h2,
+h3,
+h4,
+h5,
+h6 {
+  font-family: Karla, 'Roboto', Arial, Helvetica, sans-serif;
+  font-weight: bold;
+  color: #333;
+  margin: 0;
+}
+
+p {
+  line-height: 1.6;
+}
+
+a,
+a:visited,
+a:active,
+a:hover {
+  color: inherit;
+  text-decoration: none;
+}
+
+.app-header {
+  width: 100%;
+  height: 10rem;
+  padding: 20px;
+  display: flex;
+  flex-direction: row;
+  justify-content: space-between;
+  align-items: center;
+  box-shadow: rgba(81, 81, 81, 0.2) 0px 0px 3px 1px;
+}
+.app-header__title{
+  color: #1890ff;
+  font-family: 'Istok Web', sans-serif;
+  text-transform: uppercase;
+  font-size: 4rem;
+  margin-bottom: 0;
+}
+.app {
+  height: 100vh;
+  display: flex;
+  flex-direction: column;
+}
+.content {
+  align-content: center;
+  display: flex;
+  flex-direction: column;
+  margin: 0 auto;
+  align-items: center;
+  flex: 1 0 auto;
+}
+.introduction {
+  display: flex;
+  flex-direction: row;
+  padding: 20px;
+  align-items: center;
+  width: 80%;
+}
+.introduction-text h1 {
+  font-family: 'Karla', Arial, Helvetica, sans-serif;
+  font-weight: bold;
+  color:#1890ff;
+
+
+}
+.introduction-text p{
+  font-family: 'Istok Web', sans-serif;
+  color: #002766;
+  font-size: 1.6rem;
+}
+.introduction-img {
+  width: 100%;
+  height: auto;
+}
+.signupSection {
+  display: flex;
+  flex-direction: column;
+  margin: 30px;
+}
+.signupSection__title{
+  font-family: 'Karla', Arial, Helvetica, sans-serif;
+  font-weight: bold;
+  color:#1890ff;
+}
+.signupSection__login {
+display: block;
+}
+.app-header__box {
+  display: flex;
+  align-items:center;
+  padding: 45px 0;
+}
+.app-profileFinder {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  margin-bottom: 30px;
+}
+.app-profileFinder__title{
+  font-family: 'Karla', Arial, Helvetica, sans-serif;
+  font-weight: bold;
+  color:#1890ff;
+
+}
+.app-profileFinder p{
+  margin-top: 10px;
+    font-size: 1.6rem;
+    color: #002766;
+    font-family: 'Istok Web', sans-serif;
+}
+footer {
+  flex-shrink: 0;
+  box-shadow: rgba(81, 81, 81, 0.2) 0px 1px 6px 3px;
+  justify-content: center;
+}
+.app-footer__title {
+  font-family: 'Karla', Arial, Helvetica, sans-serif;
+  padding: 1rem 0;
+  text-align: center;
+  font-size: 2rem;
+  color: #1890ff;
+  font-weight: bold;
+}
+.app-footer__socialmedia {
+  display: flex;
+  justify-content: space-around;
+}
+.app-footer__socialmedia-item {
+  list-style: none;
+}
+main {
+  background-color: #e6f7ff;
+}
+.app-HowToGetStarted {
+  background-color: white;
+  padding: 20px;
+  box-shadow: rgba(81, 81, 81, 0.1) 0px 0px 32px 1px;
+  width: 80%;
+  border-radius: 20px;
+  margin: 20px 0;
+}
+.app-HowToGetStarted__title {
+  font-family: 'Karla', Arial, Helvetica, sans-serif;
+  padding: 1rem 0;
+  text-align: center;
+  font-weight: bold;
+  color: #1890ff;
+
+}
+.app-HowToGetStarted__list {
+  display: flex;
+}
+.app-HowToGetStarted__item {
+  list-style: none;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  padding: 10px;
+}
+.app-HowToGetStarted__desc {
+  margin-top: 10px;
+  font-size: 1.6rem;
+  color: #002766;
+  font-family: 'Istok Web', sans-serif;
+}
+.app-HowToGetStarted__desc span {
+  background-color: #bae7ff;
+}
